+++ templates/main.xml
@@ -5,6 +5,9 @@
                 <img src="/media/bomb_logo.png"></img>
             </div>
             <div class="menu">
+                <a href="#game"><button class="menu__btn">
+                    Game
+                </button></a>
                 <a href="#login"><button class="menu__btn">
                     Login
                 </button></a>
